PlayStartを作ってARページをPlayAttackにしてModeSelectのボタンのonClickを更新したよ
